+++ tests/Stub/Fruit.php
@@ -9,4 +9,17 @@ class Fruit extends Enum
     const APPLE = 'apple';
     const BANANA = 'banana';
     const CHERRY = 'cherry';
+    const EGGPLANT = 'eggplant';
+    const AUBERGINE = 'aubergine';
+
+    public static function map(): array
+    {
+        return [
+            static::APPLE => 'Apple',
+            static::BANANA => 'Banana',
+            static::CHERRY => 'Cherry',
+            static::EGGPLANT => 'Eggplant',
+            static::AUBERGINE => 'Eggplant',
+        ];
+    }
 }