+++ tests/Stub/Number.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rexlabs\Enum\Tests\Stub;
+
+use Rexlabs\Enum\Enum;
+
+class Number extends Enum
+{
+    const ONE = 1;
+    const EIGHT = 8;
+    const TEN = 10;
+    const TWENTY_FOUR = 24;
+}